Make Currency.name optional
