+++ src/components/CurrencySelector/CurrencyOptions.tsx
@@ -1,38 +1,16 @@
 import { Combobox } from "@headlessui/react";
-import { useCurrencisQuery } from "services/aws/wise";
+import { Currency } from "./types";
 import { isEmpty } from "helpers";
 
 type Props = {
   classes?: string;
-  searchText?: string;
+  currencies: Currency[];
 };
 
 export default function CurrencyOptions({
   classes = "",
-  searchText = "",
+  currencies = [],
 }: Props) {
-  const { data: currencies = [] } = useCurrencisQuery(
-    {},
-    {
-      selectFromResult({ data = [], ...rest }) {
-        return {
-          data: data.filter((c) => {
-            // check whether query matches either the currency name or any of its keywords
-            const formatQuery = searchText.toLowerCase().replace(/\s+/g, ""); // ignore spaces and casing
-            const matchesCode = c.code.toLowerCase().includes(formatQuery);
-            const matchesName = c.name
-              .toLowerCase()
-              .replace(/\s+/g, "") // ignore spaces and casing
-              .includes(formatQuery);
-
-            return matchesCode || matchesName;
-          }),
-          ...rest,
-        };
-      },
-    }
-  );
-
   return (
     <Combobox.Options
       className={`${classes} w-full bg-white dark:bg-blue-d6 shadow-lg rounded max-h-52 overflow-y-auto scroller text-base ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm`}
