+++ src/components/CurrencySelector/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./CurrencySelector";
+export type { Currency } from "./types";
