+++ src/components/CurrencySelector/types.ts
@@ -1,0 +1,3 @@
+import { WiseCurrency } from "types/aws";
+
+export type Currency = Pick<WiseCurrency, "code" | "name">;
